CXP-1313: Move Catalog DTO & QueryInterface
