+++ components/catalogs/back/src/Domain/Catalog.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Catalogs\Domain\Catalog;
+namespace Akeneo\Catalogs\Domain;
 
 /**
  * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
