+++ components/catalogs/back/src/Infrastructure/Persistence/Catalog/GetCatalogQuery.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Akeneo\Catalogs\Infrastructure\Persistence\Catalog;
 
-use Akeneo\Catalogs\Domain\Catalog\Catalog;
-use Akeneo\Catalogs\Domain\Catalog\GetCatalogQueryInterface;
+use Akeneo\Catalogs\Application\Persistence\Catalog\GetCatalogQueryInterface;
+use Akeneo\Catalogs\Domain\Catalog;
 use Doctrine\DBAL\Connection;
 use Ramsey\Uuid\Uuid;
